requirements.txt
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bleach==6.0.0
boto3==1.28.25
botocore==1.31.25
brevitas==0.8.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
comm==0.1.4
concrete-ml==1.1.0
concrete-python==1.1.0rc1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
dependencies==2.0.1
dill==0.3.7
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.93.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
fsspec==2023.6.0
h11==0.14.0
huggingface-hub==0.16.4
humanfriendly==10.0
hummingbird-ml==0.4.8
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kaggle==1.5.16
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.23.5
onnx==1.13.1
onnxconverter-common==1.13.0
onnxmltools==1.11.0
onnxoptimizer==0.3.10
onnxruntime==1.13.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.12
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.1.3
scipy==1.10.1
six==1.16.0
skl2onnx==1.12
skops==0.5.0
skorch==0.11.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.25.0
sympy==1.12
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.2.0
torch==1.13.1
torchsummary==1.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.16
uvicorn==0.21.1
wcwidth==0.2.6
webencodings==0.5.1
xgboost==1.6.2
zipp==3.16.2
